// src/app/_core/models/customer-referent-light.ts
/**
 * objet light de CustomerReferent
 */
export interface CustomerReferentLight {
  /**
   * nom du référent
   */
  userName: string;
  /**
   * id du role du référent
   */
  roleId: number;

/**
   * Email du référent
   */
  email: string;

  /**
   * prénom
   */
  firstName:string;

  /**
   * Nom
   */
  lastName: string;

  ftUniversalId: string;

  parnasseEmail: string;
}