// src/app/_core/models/models.ts
export interface CartItemVO {
  deletingDate: string;
  id: number;
  product: ProductVO;
  cartId: number;
  addedById: number;
  modifiedById: number;
  quantity: number;
  ordinal: number;
  addingDate: Date;
  unitPriceHt: number;
  discount: number;
  discountLabel: string;
  tauxTVA: string;
  comment: string;
  commentMandatory: boolean;
  productLabel: string;
  productBillLabel: string;
  productCategory: string;
  unreferencedProductCategoryId: number;
  serial: string;
  warrantyValue: number;
  warrantyLabel: string;
  articleClassLabel: string;
  articleClassId: number;
  refDataArticleClass: ReferenceDataVO;
  proactiveSale: boolean;
  isCommentDisplayOnBill: boolean;
  customerParkItemId: number;
  nomenclature: NomenclatureVO;
  acquisitionPrice: number;
  acquisitionPriceReal: number;
  margin: number;
  marginReal: number;
  uniqueId: string;
}
// ...
export interface ChildVO {
  id: number;
  firstName: string;
  liveAtHome: boolean;
  birthYear: Date;
  customerId: number;
  complement: string;
}
// ...
export interface ConfigurationModuleVO {
  id: number;
  label: string;
  description: string;
  module: ModuleVO;
  configurationModuleItems: ConfigurationModuleItemVO[];
  moduleGroup: ModuleGroupVO;
}
// ...
export interface CustomerReferentVO {
  id: number;
  customerId: number;
  userId: number;
  userName: string;
  roleId: number;
  roleName: string;
  lastName: string;
  firstName: string;
  currentAbsence: AbsenceLight;
  email: string;
  parnasseEmail: string;
}

export interface CustomerVO {
  personCustomerRolesLight: any[];
  id: number;
  companyCustomerId: number;
  customerAffiliates: CustomerVO[];
  drakkarIdentifier: string;
  status: number;
  categoryCustomer: string;
  creationDate: Date;
  statusChangeDate: Date;
  prospectingAdmissionDate: Date;
  nicheAdmissionDate: Date;
  dateHomologation: Date;
  cancellationDate: Date;
  prospectionKoDate: Date;
  adhesionStep: string;
  fieldOfActivity: string;
  siret: string;
  coachRefId: number;
  coachRefName: string;
  membreDejaRencontre: boolean;
  eligibilite: boolean;
  compteRenduEnvoye: boolean;
  propositionAdhesionEnvoye: boolean;
  clientSigne: boolean;
  referents: CustomerReferentVO[];
  personCustomerRoles: PersonCustomerRoleVO[];
  persons: PersonVO[];
  createdById: number;
  interlocutor: string;
  statusChangedBy: number;
  context: string;
  lastContact: Date;
  customerParkItems: CustomerParkItemVO[];
  recoveryProfil: string;
  recoveryComment: string;
  recoveryExtensionAllowed: boolean;
  recoveryExtensionComment: string;
  contractCessionDate: Date;
  serviceAccess: string;
  refMembershipReason: ReferenceDataVO;
  refProspectionKoReason: ReferenceDataVO;
  refCancellationReason: ReferenceDataVO;
  freeManagement: string;
  newsletter: boolean;
  statutVente: ReferenceDataVO;
  commentaireVente: string;
  isComplaint: boolean;
  complaintValue: string;
  dateDebutComplaint: Date;
  dateFinComplaint: Date;
  isSensible: boolean;
  refJobPosition: ReferenceDataVO;
  children: ChildVO[];
  notHomologatedReasons: ReferenceDataVO[];
  acceptanceDate: Date;
  paymentDate: Date;
  isMemberValidated: boolean;
  nicheIdentifier: string;
  commiteeOfSelectionComment: string;
  dateOfTechnicalStudy: Date;
  selectionCommitteeDecision: boolean;
  installationDate: Date;
  realisationStatus: string;
  debtFlag: boolean;
  debtManualRecovery: boolean;
  debtDelay: boolean;
  debtDelayComment: string;
  customerHardwareParkItems: CustomerHardwareParkItemVO[];
  selectionCommitteeDate: Date;
  acquisitionCanals: AcquisitionCanalLight[];
  companyCustomerLastNameFirstName: string;
  companyCustomerFullName: string;
  companyCustomerRealName: string;
  companyCustomerFirstName: string;
  companyCustomerLastName: string;
  customerCrmName: string;
  customerFullName: string;
  counter: number;
  offerLabel: string;
  offresService: CustomerHardwareParkItemVO[];
  prestations: CustomerHardwareParkItemVO[];
  materiels: CustomerHardwareParkItemVO[];
  idsActualBeneficiaries: number[];
  companyCustomerCrmName: string;
  offerIds: number[];
}
// ...
export interface NicheEntityVO {
  id: number;
  name: string;
  myNicheId: number;
  active: boolean;
  dateUpdate: Date;
}

export interface NodeAnswerVO {
  id: number;
  value: string;
  label: string;
  inputId: number;
  outputId: number;
  procedures: ProcedureVO[];
}
// ...
export interface PersonCustomerRoleVO {
  id: number;
  customerId: number;
  personId: number;
  refRole: ReferenceDataVO;
  allowedToCall: boolean;
}

export interface PersonVO {
  id: number;
  categoryPerson: ReferenceDataVO;
  refSocialTitle: ReferenceDataVO;
  refFavoriteLanguage: ReferenceDataVO;
  refLocation: ReferenceDataVO;
  personRoleLabel : string;
  refSocioEconomicGroup: ReferenceDataVO;
  title: string;
  firstName: string;
  lastName: string;
  nickname: string;
  birthdate: Date;
  jobTitle: string;
  login: string;
  password: string;
  secondLogin: string;
  secondPassword: string;
  firstEventDate: Date;
  sesame: string;
  isFounderMembership: boolean;
  companyName: string;
  crmName: string;
  isHusbandWife: boolean;
  isAssistant: boolean;
  activityComplement: string;
  refCompanyType: ReferenceDataVO;
  isRetired: boolean;
  refNationality: ReferenceDataVO;
  photo: DocumentVO;
  refFavoriteContactMethod1: ReferenceDataVO;
  refFavoriteContactMethod2: ReferenceDataVO;
  refFavoriteContactMethod3: ReferenceDataVO;
  refFAI: ReferenceDataVO;
  refWifiHome: ReferenceDataVO;
  ref3GHome: ReferenceDataVO;
  expectations: string;
  digitalUses: string;
  portraitByCoach: string;
  refGlobalSatisfaction: ReferenceDataVO;
  isPresentToEvent: boolean;
  refEventAbsenceReason: ReferenceDataVO;
  refProfessionalUse: ReferenceDataVO;
  refPersonalUse: ReferenceDataVO;
  refTravelerProfile: ReferenceDataVO;
  refTechnophile: ReferenceDataVO;
  refConnected: ReferenceDataVO;
  refAssistance: ReferenceDataVO;
  sensitiveMemberComent: string;
  partnerProvideurContributorComent: string;
  stopMemberComent: string;
  behaviorWithParnasseComent: string;
  refSpender: ReferenceDataVO;
  refAccessibility: ReferenceDataVO;
  refInaccessibilityReason: ReferenceDataVO;
  refInfluence: ReferenceDataVO;
  hasConciergeService: string;
  conciergeServiceName: string;
  refNoParnasseConciergeReason: ReferenceDataVO;
  hasHomeStaff: string;
  refMembershipReason2: ReferenceDataVO;
  isHonoraryMember: boolean;
  refTestAccount: ReferenceDataVO;
  refJob: ReferenceDataVO;
  numberOfPersonAtHome: number;
  refPotentialCooptation: ReferenceDataVO;
  refPotentialTurnover: ReferenceDataVO;
  cityResidence: string;
  refCountryPrimaryResidence: ReferenceDataVO;
  refCountrySecondaryResidence1: ReferenceDataVO;
  refCountrySecondaryResidence2: ReferenceDataVO;
  refResidenceCriterion: ReferenceDataVO;
  hasMac: boolean;
  hasComputer: boolean;
  listRefDataPerson: ReferenceDataVO[];
  personTypes: ReferenceDataVO[];
  personCustomerRoles: PersonCustomerRoleVO[];
  solicitationsTarget: SolicitationTargetVO[];
  contactMethods: ContactMethodVO[];
  customerParkItemsAsUser: CustomerParkItemVO[];
  refConjointIsPresent: ReferenceDataVO;
  conjointInformations: string;
  refParnasseKnowledge: ReferenceDataVO;
  additionalInformations: string;
  membershipReasonComment: string;
  obsType: ReferenceDataVO;
  obscomment: string;
  refPersonalNetwork: ReferenceDataVO;
  refPublicNotoriety: ReferenceDataVO;
  refDigitalInfluence: ReferenceDataVO;
  isBusinessProvider: boolean;
  isSupplier: boolean;
  isSensitiveMember: boolean;
  isPartnerProvideurContributor: boolean;
  isStopMember: boolean;
  isBehaviorWithParnasse: boolean;
  refMemberColor: ReferenceDataVO;
  personAffiliateRoles: PersonAffiliateRoleVO[];
  fullName: string;
  coment: string;
  serviceName: string;
}
// ...
export interface ProcedureVO {
  id: number;
  name: string;
  repeatable: boolean;
}
// ...
export interface ReferenceDataVO {
  id: number;
  label: string;
  ordinal: number;
  active: boolean;
  dateUpdate: Date;
  parent: ReferenceDataVO;
  list: ReferenceListVO;
  children: ReferenceDataVO[];
  logoName: string;
  labelComplement: string;
  key: string;
}

export interface ReferenceListVO {
  id: number;
  name: string;
  typeData: string;
  typeList: string;
  dateUpdate: Date;
  values: string;
  idMyNiche: number;
  depth: number;
}
// ...
export interface RequestReferentVO {
  id: number;
  requestId: number;
  requestHolderUserId: number;
  userId: number;
  userName: string;
  roleId: number;
  roleName: string;
  roleDisplayName: string;
  roleInputUserId: number;
}

export interface RequestSavVO {
  requestId: number;
  cartId: number;
  createdByName: string;
  generatedAt: Date;
  attachedUserName: string;
  smmUserName: string;
  customerId: number;
  customerName: string;
  customerOfferLabel: string;
  createdAt: Date;
  title: string;
  status: string;
  cartStatus: string;
  conclusionType: string;
  caseNumber: string;
}

export interface RequestTypeVO {
  id: number;
  idFamily: number;
  label: string;
  active: boolean;
  ordinal: number;
  automatic: boolean;
  key: string;
  workflow: WorkflowVO;
  technicalKey: string;
}

export interface RequestVO {
  id: number;
  title: string;
  description: string;
  createdAt: Date;
  createdById: number;
  createdByName: string;
  modifiedAt: Date;
  modifiedById: number;
  modifiedByName: string;
  customerId: number;
  customerName: string;
  firstNameRespAffaire : string;
  lastNameRespAffaire : string
  firstNameRespCommerciale : string;
	lastNameRespCommerciale :string;
  customerNicheIdentifier: string;
  customerStatus: number;
  conclusionType: string;
  status: string;
  step: number;
  closureComment: string;
  refusalReason: string;
  refusalText: string;
  refusalDate: Date;
  generatedAt: Date;
  originType: string;
  attachedUserId: number;
  attachedUserName: string;
  realEndDate: Date;
  customerLastNameFirstName: string;
  customerDeskerFullname: string;
  isConnectedOnWf: boolean;
  requestType: RequestTypeVO;
  workflow: WorkflowVO;
  proceduresAtCreation: ProcedureVO[];
  referents: RequestReferentVO[];
  universes: ReferenceDataVO[];
  nodeAnswerList: NodeAnswerVO[];
  customersBenefList: CustomerVO[];
  accountManagerUserId: number;
  accountManagerName: string;
  accountManagerRoleId: number;
  orderScoring: number;
  closingDate: Date;
  customerOfferLabel: string;
  cartStatus: string;
  listTaskAssignedIdByRequest: number[];
  cartId: number;
  cartItems: CartItemVO[];
  cartTotalHt: number;
  cartColor: string;
  countTasksAssigned: number;
  hasLateTasks: boolean;
  customerOfferStatus: string;
  campaignId: number;
  isHolder: boolean;
  dateStatusOrder: string;
  orderStatus: string;
  interlocuteur: PersonVO;
  recepientName: string;
  recepientId: number;
  recipientNicheId: string;
}

export interface RequestAnswersVO {
  id: number;
  idRequest: number;
  answer: string;
  type : string;
  label : string;
  idFormField : string;
}

export interface RoleVO {
  id: number;
  name: string;
  idMyNiche: number;
  description: string;
  active: boolean;
  dateUpdate: Date;
  configurationModules: ConfigurationModuleVO[];
  inputUser: UserVO;
  hasReferent: boolean;
  displayName: string;
  requestReferentAutoSave: boolean;
  parentId: number;
}
// ...
export interface UserVO {
  id: number;
  ftUniversalId: string;
  email: string;
  phoneNumber: string;
  parnasseEmail: string;
  parnassePhoneNumber: string;
  active: boolean;
  dateUpdate: Date;
  name: string,
  firstName: string;
  lastName: string;
  roleNamesAsStringList: string[];
  roleNamesAsString: string;
  buildVersion: string;
  environmentName: string;
  roles: RoleVO[];
  entities: NicheEntityVO[];
  roleIds: number[];
  myBackupsId: number[];
  isMobile: boolean;
  isBlocked: boolean;
  isBoMembership: boolean;
  canalType: ReferenceDataVO;
  canal: ReferenceDataVO;
  acquisitionChannelComplement: string;
  useChorniche: boolean;
  phoneLogin: string;
  treatmentGroup: string;
  phoneNum: string;
  lastConnection: Date;
  selectedRole: RoleVO;
}
// ...
export interface WorkflowVO {
  id: number;
  isSwitchOff: boolean;
}
// ...
export interface AcquisitionCanalLight {
  id: number;
  customerId: number;
  canalKey: string;
  canalName: string;
  typeKey: string;
  typeName: string;
  details1: string;
  details2: string;
  details3: string;
  customerOtherId: number;
  customerOtherFullName: string;
  userId: number;
  numberCanal: number;
  otherUserId: number;
}